extract: logic from DefineContentStep
All the logic has been extracted into its context store.
The component has been divided into constituent subcomponents -
heading, dropdowns, explanation.

The store has been placed above the template base modal.
State relevant for this step is therefore kept all the time
as a user navigates between the steps
without having to initialize it as before.
This applies also to other steps in the workflow.
